Got the modal showing
