+++ src/MoreSpeakers.Web/Pages/Mentorship/Requests.cshtml
@@ -18,7 +18,7 @@
             <h2>Mentorship Requests</h2>
             <p class="text-muted mb-0">Manage your incoming and outgoing mentorship requests</p>
         </div>
-        <a href="/Mentorship/Browse" class="btn btn-primary">
+        <a asp-page="/Mentorship/Browse" class="btn btn-primary">
             <i class="bi bi-search me-2"></i>Find Mentors
         </a>
     </div>
