+++ src/MoreSpeakers.Web/Views/Mentorship/_MentorResults.cshtml
@@ -111,7 +111,8 @@ else
                             <button class="btn btn-primary"
                                     hx-get="/Mentorship/Browse?handler=RequestModal&mentorId=@mentor.Id&type=@Model.MentorshipType"
                                     hx-target="#modal-container"
-                                    hx-swap="innerHTML">
+                                    hx-swap="innerHTML"
+                                    >
                                 <i class="bi bi-@(Model.MentorshipType == MentorshipType.NewToExperienced ? "chat-dots" : "people") me-2"></i>
                                 @(Model.MentorshipType == MentorshipType.NewToExperienced ? "Request Mentorship" : "Connect as Peer")
                             </button>
